Add hook for add_resource_route
