+++ railties/lib/rails/generators.rb
@@ -57,6 +57,7 @@ module Rails
         :orm => false,
         :performance_tool => nil,
         :resource_controller => :controller,
+        :resource_route => true,
         :scaffold_controller => :scaffold_controller,
         :stylesheets => true,
         :stylesheet_engine => :css,
@@ -182,6 +183,7 @@ module Rails
 
         [
           "rails",
+          "resource_route",
           "#{orm}:migration",
           "#{orm}:model",
           "#{orm}:observer",
