+++ railties/lib/rails/generators/rails/resource/resource_generator.rb
@@ -14,13 +14,7 @@ module Rails
       class_option :actions, :type => :array, :banner => "ACTION ACTION", :default => [],
                              :desc => "Actions for the resource controller"
 
-      def add_resource_route
-        return if options[:actions].present?
-        route_config =  regular_class_path.collect{|namespace| "namespace :#{namespace} do " }.join(" ")
-        route_config << "resources :#{file_name.pluralize}"
-        route_config << " end" * regular_class_path.size
-        route route_config
-      end
+      hook_for :resource_route, :required => true
     end
   end
 end
